+++ assets/css/styles.css
@@ -1,0 +1,118 @@
+:root {
+    --color-background: #121212 ;
+    --color-background-2: #202020;
+    --color-text: #FFFFFF;
+    --color-text-2: #BBBBBB;
+    --color-box:#FFFFFF ;
+    --color-box-2: #BBBBBB;
+    --accent-color: #3F51B5;
+    --accent-color-2: #ff4081;
+}
+
+html {
+    color: var(--color-text);
+    background-color: var(--color-background);
+    text-align: center;
+    font-family: "Roboto Mono", monospace;
+}
+
+body {
+    background-color: var(--color-background);
+    margin: auto;
+    color: var(--color-text);
+}
+
+h1 {
+    font-size: 3em;
+    font-weight: bolder;
+    margin: 0px;
+}
+
+header {
+    background-color: var(--color-background-2);
+    text-align: center;
+    border-radius: 0px;
+    /* display: grid;
+    grid-template-columns: 1fr 1fr 1fr 1fr; */
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding-right: 5%;
+}
+
+#main {
+    height: 100%;
+    background-color: var(--color-background);
+    min-height: 200px;
+    padding: 50px;
+    display: grid;
+    gap: 50px;
+    grid-template-columns: 1fr 1fr 1fr;
+}
+
+.content {
+    display: grid;
+    background-color: var(--color-background-2);
+    border-radius: 5%;
+    min-height: 200px;
+}
+
+.description {
+    margin: 5px;
+}
+
+footer {
+    background-color: var(--color-background-2);
+    text-align: center;
+    padding: 15px;
+    align-items: center;
+}
+
+.imgArticle {
+    border-radius: 5%;
+    border-bottom-right-radius: 0px;
+    border-bottom-left-radius: 0px;
+    width: 100%;
+}
+
+a {
+    color: var(--color-text);
+    text-decoration: none;
+}
+
+button {
+    background-color: var(--color-background-2);
+    font-family: "Roboto Mono", monospace;
+    font-size: 1em;
+    color: var(--color-text);
+    border: 1px solid var(--color-text);
+    border-radius: 10%;
+    padding: 5px;
+    margin: 5px;
+}
+button:hover {
+    background-color: var(--color-text);
+    color: var(--color-background-2);
+    transition: 0.3s;
+}
+
+.imglogo {
+    height: 100%;
+}
+
+
+/*---------------------------------------------*/
+/* New Part*/
+
+.item-list {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    gap: 20px;
+  }
+  
+  .item {
+    border: 1px solid #ccc;
+    padding: 10px;
+    text-align: center;
+  }
+
