+++ assets/json/items.json
@@ -1,0 +1,17 @@
+[
+    {
+      "id": 1,
+      "name": "Item 1",
+      "price": 100,
+      "image": "assets/images/item1.jpg",
+      "description": "This is a great product."
+    },
+    {
+      "id": 2,
+      "name": "Item 2",
+      "price": 150,
+      "image": "assets/images/item2.jpg",
+      "description": "Another amazing product."
+    }
+  ]
+
