+++ index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="fr">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>Vente Au Cess</title>
+        <link rel="stylesheet" href="assets/css/styles.css" />
+        <script src="assets/js/script.js"></script>
+    </head>
+    <body>
+        <header>
+            <h1>My Online Store</h1>
+        </header>
+        <main>
+            <section class="item-list">
+                <!-- List items will go here -->
+            </section>
+        </main>
+        <footer>
+            <p>© 2024 Vente Au Cess</p>
+        </footer>
+    </body>
+</html>
