+++ item.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>Item Details</title>
+        <link rel="stylesheet" href="assets/css/styles.css" />
+    </head>
+    <body>
+        <header>
+            <h1>Item Details</h1>
+            <a href="index.html">Back to Items</a>
+        </header>
+        <main>
+            <div class="item-details">
+                <!-- Dynamic content will go here -->
+            </div>
+        </main>
+        <script src="assets/js/item.js"></script>
+    </body>
+</html>
